Create AI manager and restructure data
