+++ js/managers/MetaAiManager.js
@@ -1,0 +1,9 @@
+export class MetaAiManager {
+    constructor() {
+        this.managers = {};
+    }
+
+    addManager(name, manager) {
+        this.managers[name] = manager;
+    }
+}
