+++ js/managers/index.js
@@ -3,3 +3,5 @@ export { VisualEffectManager } from './VisualEffectManager.js';
 export { EventManager } from './EventManager.js';
 export { StatusEffectManager } from './StatusEffectManager.js';
 export { battleMaster } from './BattleMaster.js';
+export { AiManager } from './AiManager.js';
+export { MetaAiManager } from './MetaAiManager.js';
